https://bdslearningapp.blogspot.com/2020/08/electric-charges-and-fields.html
In 600 B.C., Thales of Miletus, one of the founders of Greek science, first noticed that if a piece of amber is rubbed with a woolen cloth, it then acquires the property of attracting light feathers, dust, lint, pieces of leaves, etc.
In 1600 A.D., William Gilbert, the personal doctor to Queen Elizabeth- I of England, made a systematic study of the substance that behaves like amber. In his book De Magnete (on the magnet), he introduces the name electrica for such substance. In fact, the Greek name for amber is an electron, which is the origin of all such words: electricity, electric force, electric charge, and electron.